Add verbose (#3)
Add verbose mode.

To improve the logs I've added a verbose flag to the application. With the verbose flag it will print all the actions being done in almost every step, and for the commands it will redirect the process output to the application output. 
The command parsing was improved to allow Windows paths in the YAML.